Implementacion de seguridad parte 1

+++ src/main/java/com/example/demo/usuario/domain/UsuarioService.java
@@ -3,7 +3,7 @@ package com.example.demo.usuario.domain;
 import com.example.demo.limite.domain.Limite;
 import com.example.demo.limite.infraestructure.LimiteRepository;
 import com.example.demo.usuario.dto.UsuarioDto;
-import com.example.demo.usuario.infraestructure.SolicitudRepository;
+import com.example.demo.solicitud.infrastructure.SolicitudRepository;
 import com.example.demo.usuario.infraestructure.UsuarioRepository;
 import org.springframework.stereotype.Service;
 
@@ -57,7 +57,7 @@ public class UsuarioService {
         Usuario usuario = usuarioRepository.findById(usuarioId)
                 .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
 
-        int totalSolicitudes = solicitudRepository.countByUsuarioId(usuarioId);
+        int totalSolicitudes = Math.toIntExact(solicitudRepository.countByUsuarioId(usuarioId));
         int totalTokens = solicitudRepository.sumTokensByUsuarioId(usuarioId);
 
         return new UsuarioDto(usuarioId, totalSolicitudes, totalTokens);
